moves the nav to the blocks, selects a separate intro section from the header, adds a pseudo-element (line) to the nav
