+++ src/components/blocks/consultations/consultations.jsx
@@ -5,13 +5,13 @@ import { TitleLevel } from '../../../utils/consts';
 // import styles from './style.module.css';
 
 function Consultations() {
-  return(
-  <section>
+  return (
+    <section className='wrapper'>
       <Container>
         <Title level={TitleLevel.H2}>С какими запросами я могу помочь:</Title>
       </Container>
     </section>
-    );
+  );
 }
 
 export default Consultations;
