+++ src/components/blocks/intro/style.module.css
@@ -1,0 +1,32 @@
+.intro__wrapper {
+  display: flex;
+  justify-content: space-around;
+  flex-wrap: wrap;
+  padding: 2em 0;
+}
+
+.intro__content {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  max-width: 50%;
+}
+
+.intro__content h2 {
+  font-weight: 400;
+  font-size: 2.5rem;
+}
+
+.intro__img {
+  max-width: 35%;
+  overflow: hidden;
+}
+
+.intro__img img {
+  width: 100%;
+  height: 100%;
+  border-radius: 15px;
+  object-fit: cover;
+
+  /* margin-right: 200px; */
+}
