+++ src/components/layout/header/style.module.css
@@ -1,39 +1,6 @@
-.header__top {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 40px 0;
-}
-
 .header__wrapper {
   display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  padding: 2em 0;
-}
-
-.header__content {
-  display: flex;
-  flex-direction: column;
+  align-items: center;
   justify-content: center;
-}
-
-.header__content h2 {
-  font-weight: 400;
-  font-size: 2.5rem;
-}
-
-.header__img {
-  width: 350px;
-  height: 350px;
-  overflow: hidden;
-}
-
-.header__img img {
-  width: 100%;
-  height: 100%;
-  border-radius: 15px;
-  object-fit: cover;
-
-  /* margin-right: 200px; */
+  padding: 40px 0;
 }
